requirements.txt
gymnasium>=0.29
mujoco>=3.2
imageio>=2.34
matplotlib>=3.8
numpy>=1.26
scipy>=1.11
yamlreader>=3.0
stable-baselines3
torch
tensorboard
tqdm
imageio[ffmpeg]